Fix brokerstats

+++ pkg/types/broker_stats.go
@@ -108,7 +108,7 @@ type BrokerLoadStats struct {
 	BrokerState        BrokerState `json:"BrokerState"`
 	Broker             int32       `json:"Broker"`
 	NwOutRate          float64     `json:"NwOutRate"`
-	NumCore            int32       `json:"NumCore"`
+	NumCore            float64     `json:"NumCore"`
 	Host               string      `json:"Host"`
 	CPUPct             float64     `json:"CpuPct"`
 	Replicas           int32       `json:"Replicas"`
